Add Placeholders when lists are empty

+++ lazybear/Views/Watchlist.swift
@@ -15,19 +15,26 @@ struct Watchlist: View {
     
     var body: some View {
         NavigationView {
-            List {
-                Section(header: Text("Watchlist")) {
-                    ForEach(companies) { company in
-                        WatchlistRow(company: company)
+            VStack {
+                if companies.count > 0 {
+                    List {
+                        Section(header: Text("Watchlist")) {
+                            ForEach(companies) { company in
+                                WatchlistRow(company: company)
+                            }
+                            .onDelete { indexSet in deleteWatchlist(indexSet: indexSet) }  // Delete from persistent storage
+                        }
                     }
-                    .onDelete { indexSet in deleteWatchlist(indexSet: indexSet) }  // Delete from persistent storage
+                    .toolbar { EditButton() }
+                } else {
+                    Placeholder(title: "Seems that your watchlist is empty!", text: "Search and add a company.")
                 }
             }
             .navigationBarTitle("Watchlist", displayMode: .inline)
-            .toolbar { EditButton() }
-            
-        }.navigationViewStyle(StackNavigationViewStyle())
+        }
+        .navigationViewStyle(StackNavigationViewStyle())
     }
+  
     
     func deleteWatchlist(indexSet: IndexSet) {
         for index in indexSet {
